requirements.txt
# --- CORE DJANGO & API ---
Django>=5.0.1,<5.1
djangorestframework>=3.15.1
django-cors-headers>=4.3.1
django-filter>=24.1
drf-spectacular>=0.27.1

# --- DATABASE & CACHE ---
psycopg2-binary>=2.9.9
pgvector>=0.2.4
django-redis>=5.4.0

# --- AUTH & USERS ---
djangorestframework-simplejwt>=5.3.1
django-allauth>=0.61.1

# --- CELERY & REDIS ---
celery>=5.3.6
redis>=5.0.1
django-celery-beat>=2.6.0
django-celery-results>=2.5.1

# --- MACHINE LEARNING ---
torch>=2.2.0
sentence-transformers>=2.5.1
scikit-learn>=1.4.0
numpy>=1.26.3

# --- WEBSOCKETS & ASGI ---
channels>=4.0.0
daphne>=4.0.0

# --- UTILS & SCRAPING ---
httpx>=0.26.0
beautifulsoup4>=4.12.3
feedparser>=6.0.11
fuzzywuzzy>=0.18.0
python-Levenshtein
python-dotenv>=1.0.1
pillow>=10.2.0
python-dateutil>=2.8.2
django-prometheus>=2.3.1
cryptography>=42.0.5